+++ src/main/java/com/igortauh/crudpoo/repositories/VeiculoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 }
